backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
openai==1.35.3
pdfplumber==0.11.0
PyMuPDF==1.24.5
python-pptx==0.6.23
python-multipart==0.0.9
pydantic[email]==2.7.4
